Add ability to use Access-Control-Allow-Origin
Ran into issues when running mock-yeah locally that didn't allow for the use of
provided fixture data. This was an issue we've run into before, and we solved it
via middleware, let me know if you think there is a better way of doing this.

I added a config value `accessControlAllowOrigin`, that is set to `false` by default.
Setting this to true will allow the app to use the Access-Control-Allow-Origin middleware.
This should allow users to run mock-yeah locally on projects that have configurations
that would cause this error.